Rename pilight base entity module
