+++ homeassistant/components/pilight/entity.py
