+++ homeassistant/components/pilight/light.py
@@ -18,8 +18,8 @@ import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from .base_class import SWITCHES_SCHEMA, PilightBaseDevice
 from .const import CONF_DIMLEVEL_MAX, CONF_DIMLEVEL_MIN
+from .entity import SWITCHES_SCHEMA, PilightBaseDevice
 
 LIGHTS_SCHEMA = SWITCHES_SCHEMA.extend(
     {
